klringbuffer: destroy mutex on free, check realloc(), consistent NULL/size_t
- rb_free() now unlocks before pthread_mutex_destroy() (destroying a
  locked mutex is UB) and calls it unconditionally, matching rb_new()'s
  unconditional pthread_mutex_init() -- every ring previously leaked its
  mutex's resources and could free the struct while still holding its
  own lock.
- _rb_shrink_reset() now checks realloc()'s result. On failure it keeps
  the existing (larger than ideal, but valid) buffer instead of
  overwriting buf->data with NULL, which leaked the original allocation
  and handed the next write a NULL pointer.
- rb_get_write_pos()/rb_get_read_pos() now return size_t, matching
  every other size-related accessor in the file, instead of
  unsigned int (a truncation risk on 64-bit builds for large rings).
  Added their missing doc comments.
- rb_new() returns NULL instead of the literal 0 on failure.

Verified clean under -fsanitize=address,undefined; test_klringbuffer
and test_pes_extractor (the other consumer, which narrows the new
size_t position values into its own unsigned int ringPos convention
unchanged) both build and pass.

+++ src/klringbuffer.h
@@ -189,7 +189,18 @@ void rb_free(KLRingBuffer *buf);
  */
 void rb_discard(KLRingBuffer *buf, size_t bytes);
 
-unsigned int rb_get_write_pos(KLRingBuffer *buf);
-unsigned int rb_get_read_pos(KLRingBuffer *buf);
+/**
+ * @brief       Current write offset within the ring's physical allocation.
+ * @param[in]   KLRingBuffer *buf - Object.
+ * @return      Offset in bytes ((head + fill) % size). 0 if buf is NULL.
+ */
+size_t rb_get_write_pos(KLRingBuffer *buf);
+
+/**
+ * @brief       Current read offset (head) within the ring's physical allocation.
+ * @param[in]   KLRingBuffer *buf - Object.
+ * @return      Offset in bytes. 0 if buf is NULL.
+ */
+size_t rb_get_read_pos(KLRingBuffer *buf);
 
 #endif /* KLRINGBUFFER_H */
